edited

+++ src/ft_length.c
@@ -6,7 +6,7 @@
 /*   By: mhwangbo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/04/21 00:10:14 by mhwangbo          #+#    #+#             */
-/*   Updated: 2018/04/23 15:42:28 by mhwangbo         ###   ########.fr       */
+/*   Updated: 2018/04/23 16:12:12 by mhwangbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int				ft_d_cv(t_flag flags, va_list args)
 	return (i);
 }
 
-unsigned		ft_un_cv(t_flag flags, va_list args)
+unsigned long long		ft_un_cv(t_flag flags, va_list args)
 {
 	if (flags.length == 0 || flags.length == 1 || flags.length == 2)
 		return(va_arg(args, unsigned int));
